Cleaned up code and solved bug!
Managed to narrow down what was causing the "off-axis" bug that occurred when DLSS is active. It's to do with SetPipelineResolution, if it isn't already set to the DLSS width and Height, it wouldn't apply correctly!

The bug still occurs at other resolutions, but atleast it no longer occurs on the starting one :)

Also began implementing DLSS modes, but these are not working just yet. Will continue to clean up the off-axis bug, and then complete the mode feature in-tandem.